tests: move tenant_foobar fixture to common fixtures module
The fixture can be reused by other tests

Changelog: None
